Refactor sensible.zsh by removing XDG cache setup and update install_lazyvim.sh to prompt for configuration removal. Enhance .zshrc with additional plugin management and key binding for command editing.

+++ lazyvim/install_lazyvim.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
 
-rm -rf ~/.config/nvim
-rm -rf ~/.local/share/nvim
-rm -rf ~/.local/state/nvim
-rm -rf ~/.cache/nvim
+echo "Do you want to remove existing Neovim configurations? (y/n)"
+read -r response
+if [[ "$response" == "y" ]]; then
+    rm -rf ~/.config/nvim
+    rm -rf ~/.local/share/nvim
+    rm -rf ~/.local/state/nvim
+    rm -rf ~/.cache/nvim
+fi
 
 git clone https://github.com/LazyVim/starter ~/.config/nvim
 rm -rf ~/.config/nvim/.git
